feat(get): get rds command : move pkg/getRdsInfo to pkg/rds/rds.go
removed pkg/getRdsInfo.go and created pkg/rds/rds.go
